trying to match types

+++ leetcode/src/remove_element.rs
@@ -2,7 +2,7 @@ pub fn remove_element(nums: &mut Vec<i32>, val: i32) -> i32 {
     nums.sort();
     for n in nums {
         if n != &val {
-nums.remove(n as u32);
+nums.remove(n);
         }
 
     }
